distributed-llm-alignment/requirements.txt
torch
transformers
datasets
accelerate
deepspeed
sentencepiece
bitsandbytes
peft
evaluate
wandb
tqdm
pyyaml
scipy
numpy
rich
